requirements.txt
uvicorn~=0.18.2
fastapi~=0.79.0
pydantic~=1.9.2
SQLAlchemy~=1.4.40

PyJWT==1.7.1
python-decouple==3.3
